% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Implements agglomerative hierarchical clustering}
\usage{
hier_clust(data, k)
}
\arguments{
\item{k}{Number of clusters}

\item{dat}{A dataframe}
}
\value{
A data frame of cluster assignments (merge and order?)
}
\description{
Implements agglomerative hierarchical clustering
}
